Moved out main function to separate file.
